refactor(auth): mail service
